+++ JavaSchedulingSystem/nbproject/project.properties
@@ -78,7 +78,7 @@ javafx.enable.concurrent.external.runs=false
 javafx.enabled=true
 javafx.fallback.class=com.javafx.main.NoJavaFXFallback
 # Main class for JavaFX
-javafx.main.class=javaschedulingsystem.JavaSchedulingSystem
+javafx.main.class=main.JavaSchedulingSystem
 javafx.preloader.class=
 # This project does not use Preloader
 javafx.preloader.enabled=false
