+++ JavaSchedulingSystem/src/view_controller/AddAppointmentView.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.text.Font?>
 
 
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="javaschedulingsystem.AddAppointmentViewController">
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="view_controller.AddAppointmentViewController">
    <children>
       <Label layoutX="188.0" layoutY="152.0" text="Date:" />
       <Label layoutX="33.0" layoutY="33.0" text="Add New Appointment">
