+++ JavaSchedulingSystem/src/view_controller/AddCustomerView.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.Font?>
 
 
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="javaschedulingsystem.AddCustomerViewController">
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="view_controller.AddCustomerViewController">
    <children>
       <Label layoutX="62.0" layoutY="35.0" text="Add New Customer">
          <font>
