+++ JavaSchedulingSystem/src/view_controller/EditDeleteAppointmentsView.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.text.Font?>
 
 
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="javaschedulingsystem.EditDeleteAppointmentsViewController">
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="view_controller.EditDeleteAppointmentsViewController">
    <children>
       <Label layoutX="52.0" layoutY="38.0" text="Edit or Delete Appointment">
          <font>
