+++ JavaSchedulingSystem/src/view_controller/EditDeleteCustomerView.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.Font?>
 
 
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="javaschedulingsystem.EditDeleteCustomerViewController">
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="view_controller.EditDeleteCustomerViewController">
    <children>
       <Label layoutX="44.0" layoutY="32.0" text="Edit or Delete Customer">
          <font>
