+++ JavaSchedulingSystem/src/view_controller/LoginView.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="javaschedulingsystem.LoginViewController">
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view_controller.LoginViewController">
    <children>
       <PasswordField id="loginPassword" fx:id="passwordField" layoutX="320.0" layoutY="248.0" promptText="Password" />
       <TextField id="loginUsername" fx:id="usernameField" layoutX="320.0" layoutY="200.0" promptText="Username" />
