+++ JavaSchedulingSystem/src/view_controller/ReadOnlyCustomerView.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="javaschedulingsystem.ReadOnlyCustomerViewController">
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view_controller.ReadOnlyCustomerViewController">
    <children>
       <Label layoutX="42.0" layoutY="35.0" text="Customer Details">
          <font>
