+++ JavaSchedulingSystem/src/view_controller/SpecificReportView.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 
 
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="javaschedulingsystem.SpecificReportViewController">
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="800.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="view_controller.SpecificReportViewController">
    <children>
       <Button fx:id="backButton" layoutX="703.0" layoutY="347.0" mnemonicParsing="false" onAction="#backButtonHandler" text="Back" />
       <Label layoutX="54.0" layoutY="60.0" text="Specific Reports View" />
